split json and odin package in pbmm
