+++ bmm/src/main/java/org/openehr/bmm/v2/persistence/odin/BmmOdinParser.java
@@ -1,4 +1,4 @@
-package org.openehr.bmm.v2.persistence.json;
+package org.openehr.bmm.v2.persistence.odin;
 
 import com.nedap.archie.adlparser.antlr.odinLexer;
 import com.nedap.archie.adlparser.antlr.odinParser;
@@ -8,6 +8,7 @@ import com.nedap.archie.serializer.odin.OdinToJsonConverter;
 import org.antlr.v4.runtime.CharStreams;
 import org.antlr.v4.runtime.CommonTokenStream;
 import org.openehr.bmm.v2.persistence.PBmmSchema;
+import org.openehr.bmm.v2.persistence.jackson.BmmJacksonUtil;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -17,7 +18,7 @@ public class BmmOdinParser {
     public static PBmmSchema convert(odinParser.Odin_textContext odin) {
         try {
             String json = new OdinToJsonConverter().convert(odin);
-            return BmmJacksonOdinUtil.getObjectMapper().readValue(json, PBmmSchema.class);
+            return BmmJacksonUtil.getObjectMapper().readValue(json, PBmmSchema.class);
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
