+++ referencemodels/src/main/java/org/openehr/referencemodels/BuiltinReferenceModels.java
@@ -6,10 +6,8 @@ import com.nedap.archie.rminfo.MetaModels;
 import com.nedap.archie.rminfo.ModelInfoLookup;
 import com.nedap.archie.rminfo.ReferenceModels;
 import org.openehr.bmm.rmaccess.ReferenceModelAccess;
-import org.openehr.bmm.v2.persistence.PBmmSchema;
-import org.openehr.bmm.v2.persistence.json.BmmOdinParser;
+import org.openehr.bmm.v2.persistence.odin.BmmOdinParser;
 import org.openehr.bmm.v2.validation.BmmSchemaConverter;
-import org.openehr.bmm.v2.validation.BmmValidationResult;
 import org.openehr.bmm.v2.validation.BmmRepository;
 import org.reflections.Reflections;
 import org.reflections.scanners.ResourcesScanner;
